Adds a ResourceAssembler and some tests
